requirements.txt
<<<<<<< HEAD
affine==2.3.1
aiofiles==0.5.0
amqp==5.0.6
aniso8601==7.0.0
anyio==3.5.0
asgiref==3.3.4
async-exit-stack==1.0.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
contextlib2==21.6.0
contextvars==2.4
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
Django==3.2.3
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.5.0
django-mass-edit==3.4.1
django-simple-history==3.0.0
django-tables2==2.4.0
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.73.0
fontawesomefree==5.15.3
fpdf2==2.5.2
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.1.2
idna==3.2
image-similarity-measures==0.3.5
imageio==2.9.0
immutables==0.16
importlib-metadata==2.1.1
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.1.0
lpips==0.1.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.3.4
multiprocess==0.70.12.2
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.0.3
networkx==2.5.1
Nuitka==0.6.16.2
numpy==1.19.5
odfpy==1.4.1
opencv-python==4.5.2.54
openpyxl==3.0.7
orjson==3.6.0
packaging==21.0
pandas==1.1.5
passlib==1.7.4
pathos==0.2.8
phasepack==1.5
Pillow==8.2.0
pkg_resources==0.0.0
pluggy==0.13.1
pox==0.3.0
ppft==1.6.6.4
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.19.4
py==1.10.0
pydantic==1.8.2
pyFFTW==0.12.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
rasterio==1.2.10
reportlab==3.6.8
requests==2.26.0
Rx==1.6.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
sewar==0.4.5
six==1.16.0
sklearn==0.0
snakeviz==2.1.1
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.2.1
sqlparse==0.4.1
starlette==0.14.2
tablib==3.0.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.14.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
wurlitzer==3.0.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
zmq==0.0.0
=======
cryptography
Cython
Django
django-admin-list-filter-dropdown
django-admin-rangefilter
django-bootstrap-datepicker-plus
django-bootstrap4
django-cleanup
django-extensions
django-filter
django-import-export
django-mass-edit
django-simple-history
django-tables2



fontawesomefree

gunicorn


mysql-connector
mysql-connector-python
mysqlclient
numpy
opencv-python==4.5.3.56

pathos
passlib

reportlab
requests

scikit-image
scikit-learn
scipy

termcolor
threadpoolctl

uvicorn


wcwidth

wurlitzer


>>>>>>> added heap of changes that were not pushed up since september 2022.  Some known - fixed bug with pdf creation where log or reference image were deleted.  Added code to push message to synergy. Current version has Synergy skin
